Abstracted the navbar, header and footer out as a components.
