+++ app/components/header.tsx
@@ -1,6 +1,6 @@
 import React from 'react';
 
-import Navbar from './navbar';
+import Navbar from './navbar/navbar';
 
 export default function Header() {
   return (
