+++ app/components/navbar/navigationLinks.json
@@ -1,0 +1,10 @@
+[
+    {
+        "label": "Home",
+        "href": "/#"
+    },
+    {
+        "label": "Contributors",
+        "href": "/contributors"
+    }
+]
